Conflicts:
	Project/Zazo.xcodeproj/project.pbxproj
	Project/Zazo/UserStory/Main/Grid Module/ApplicationLogic/Interactor/ZZGridInteractor.m
	Project/Zazo/UserStory/Main/Grid Module/ApplicationLogic/Interactor/ZZGridInteractorIO.h
	Project/Zazo/UserStory/Main/Grid Module/UserInterface/Presenter/ZZGridPresenter.h
	Project/Zazo/UserStory/Main/Grid Module/UserInterface/Presenter/ZZGridPresenter.m

+++ Project/tbm/Logic/Categories/NSArray/NSArray+TBMArrayHelpers.m
@@ -11,7 +11,7 @@
 {
     NSUInteger myCount = [self count];
     if (myCount)
-        return [self objectAtIndex:arc4random_uniform(myCount)];
+        return [self objectAtIndex:arc4random_uniform((u_int32_t)myCount)];
     else
         return nil;
 }
